Move the `Token` class under the `classes` folder
