+++ src/plugins/diff-highlight/prism-diff-highlight.ts
@@ -1,8 +1,8 @@
-import { Token, getTextContent } from '../../core/token';
+import { Token, getTextContent } from '../../core/classes/token';
 import diff, { PREFIXES } from '../../languages/diff';
 import { addHooks } from '../../shared/hooks-util';
 import type { BeforeSanityCheckEnv, BeforeTokenizeEnv } from '../../core/hooks';
-import type { TokenStream } from '../../core/token';
+import type { TokenStream } from '../../core/classes/token';
 import type { PluginProto } from '../../types';
 
 export default {
