+++ src/shared/languages/templating.ts
@@ -1,8 +1,8 @@
-import { getTextContent } from '../../core/token';
+import { getTextContent } from '../../core/classes/token';
 import { withoutTokenize } from '../language-util';
 import type { Prism } from '../../core';
+import type { Token, TokenStream } from '../../core/classes/token';
 import type { Registry } from '../../core/registry';
-import type { Token, TokenStream } from '../../core/token';
 import type { Grammar } from '../../types';
 import type { tokenize } from '../symbols';
