+++ src/shared/tokenize-strings.ts
@@ -1,4 +1,4 @@
-import type { TokenStream } from '../core/token';
+import type { TokenStream } from '../core/classes/token';
 
 /**
  * Given a token stream and a tokenization function, this will tokenize all
